+++ crates/mongodb-support/src/aggregate/mod.rs
@@ -1,9 +1,11 @@
 mod accumulator;
 mod pipeline;
 mod selection;
+mod sort_document;
 mod stage;
 
 pub use self::accumulator::Accumulator;
 pub use self::pipeline::Pipeline;
 pub use self::selection::Selection;
+pub use self::sort_document::SortDocument;
 pub use self::stage::Stage;